here andre

+++ server_code/includes/neighbor.h
@@ -11,8 +11,8 @@ enum{
 
 typedef struct Neighbor{
     nx_uint8_t address; // Identifier for node
-    uint16_t pktReceived; // For calculations, X = total packets received
-    uint16_t pktSent; // For calculations, Y = total packets sent
+    uint8_t pktReceived; // For calculations, X = total packets received
+    uint8_t pktSent; // For calculations, Y = total packets sent
     float qol; // Calculated for quality of life (quality of link hahaha)
     bool active;
 } Neighbor;
